#Refactoring & Improvements - Create CellTypeConstants. - Create CellTypeHelper. - Create LocalPartConstants. - Create Adapters for not supported operations. - Remove duplicated code from StreamingCell. - Fix bug when retrieving Datetime with only time from Cell. - Implement getAddress method in StreamingCell. - Make sure than StreamingSheetReader use LocalPartConstants. - Make sure than BufferedStringsTable use LocalPartConstants. - Make sure than Utilities classes is final by default. - Upgrade poi version to 5.2.2.

+++ src/main/java/com/monitorjbl/xlsx/sst/LRUCache.java
@@ -39,7 +39,7 @@ class LRUCache  {
 //  just an estimation
     private static long strSize(String str) {
         long size = Integer.BYTES; // hashCode
-        size += Character.BYTES * str.length(); // characters
+        size += (long) Character.BYTES * str.length(); // characters
         return size;
     }
